Fixed Nuxt not loading in dev environment
For `docker-compose.override`, the `app` directory is installed in the image, the `node_modules` directory is installed as well. However, when run, then the `frontend` directory from outside is mounted instead, and the environment fails to run.

This fix has been applied: https://stackoverflow.com/a/32785014

NOTE: this does mean that any new `npm` packages installed will require a rebuild of the `frontend`.